refactor(git-helper): extract and move method to clone repository and branch with sanitized name ➡️ ✨
